feat: add Telegram channels as source type
Telegram channels are a new source type for Sugarcube. This
implementation allows input such as `https://t.me/s/<channel>` and
`https://t.me/<channel>`. Channel names such as `@<channel>` are not
recognized since they conflict with Twitter user names. To resolve this
I'd need to introduce an ambiguous `SourceType` type, which allows
multiple source types as a result to the `sourceType` function.